+++ src/frontend/src/modules/builder/components/BuilderIngredientsSelector.vue
