+++ src/frontend/src/modules/builder/components/BuilderPizzaView.vue
