+++ src/frontend/src/modules/builder/components/BuilderSizeSelector.vue
