+++ src/frontend/src/router/index.js
@@ -1,11 +1,11 @@
-import Vue from 'vue';
-import Router from 'vue-router';
-import routes from '@/router/routes';
+import Vue from "vue";
+import Router from "vue-router";
+import routes from "@/router/routes";
 
 Vue.use(Router);
 
 export default new Router({
   mode: "history",
   base: process.env.BASE_URL,
-  routes,
+  routes
 });
