+++ src/frontend/src/router/routes.js
@@ -1,10 +1,7 @@
 export default [
   {
     path: "/",
-    name: "IndexHome",
-    components: () => import("../views/Index"),
-    meta: {
-      layout: "AppLayoutMain",
-    },
+    name: "Index",
+    components: () => import("../views/Index.vue"),
   },
 ];
