--- src/frontend/src/views/Cart.vue
@@ -1,20 +1,0 @@
-<template>
-  <div>
-    <CartBuilder />
-  </div>
-</template>
-
-<script>
-import CartBuilder from "@/modules/components/cart/CartBuilder";
-
-export default {
-  name: "Cart",
-  components: {
-    CartBuilder,
-  },
-};
-</script>
-
-<style scoped>
-
-</style>
