+++ src/frontend/src/views/CartView.vue
@@ -1,0 +1,15 @@
+<template>
+  <div>
+    <h1>asdasdsdasdasdsa</h1>
+  </div>
+</template>
+
+<script>
+// import CartBuilder from "@/modules/components/cart/CartBuilder";
+
+export default {
+  name: "Cart",
+};
+</script>
+
+<style scoped></style>
